fix bug and update report

+++ TravelTourBooking/TravelTourBooking.BLL/Services/DestinationService.cs
@@ -1,4 +1,4 @@
-﻿using AutoMapper;
+using AutoMapper;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -52,6 +52,10 @@ namespace TravelTourBooking.BLL.Services
         {
             if (!await repo.ExistsAsync(id))
                 throw new KeyNotFoundException($"Điểm đến ID {id} không tồn tại.");
+
+            if (await repo.HasToursAsync(id))
+                throw new InvalidOperationException("Không thể xóa điểm đến này vì đang có các tour thuộc điểm đến này.");
+
             await repo.DeleteAsync(id);
         }
     }
